Move `std::io::Write` to `core::io`
